Correct the mapping and Move Customer Specification into Khatabook

+++ khatabook/src/main/java/com/generic/khatabook/model/UnitOfValue.java
@@ -2,12 +2,11 @@ package com.generic.khatabook.model;
 
 import java.math.BigDecimal;
 
-public record UnitOfValue(BigDecimal price, int start, int end) {
+public record UnitOfValue(BigDecimal price, Long start, Long end) {
 
-    public static  UnitOfValue non(){
-        return new UnitOfValue(BigDecimal.ZERO, 0, 0);
+    public static UnitOfValue non() {
+        return new UnitOfValue(BigDecimal.ZERO, 0l, 0l);
     }
 
 }
-
